functions/process-image/requirements.txt
functions-framework
google-cloud-storage
google-cloud-pubsub
pillow